refactor: introduce shared code to manage description of story
Add a new function taking an HTML content that will be used to display the description of the Story in a dedicated div.
